feat(git): add `script/makeReleaseNotes.ts`
This script will generate release notes from conventional commits since the last tag. Often the result is satisfactory as-is, and it's otherwise a great starting point.

See the readme entry or run with `--help` option for full info. It's recommended to use the `-g` flag to group commits by type, the `-c` flag to copy the output to clipboard, and the `--types` flag to filter and order the types of commits to include.

```sh
deno run -A https://deno.land/x/handy/git/script/makeReleaseNotes.ts \
  -gc --types=feat,fix
```