feat(collection): add `position` submodule
A `Position` describes a location between items in an `IndexedCollection` such as an array or string. For example `"a"` has two `Positions` - `0` and `1`, representing the locations before and after `a`. This has some unique properties: `0` is always a valid, and the collection's `length` is always valid. These semantics correspond to the behavior of the first argument of `Array.splice`.

The function `toPosition` takes a positive or negative index and normalizes it into a positive `Position` value, where `-1` refers to the position before the last item in the collection. The special value `-0` refers to the `Position` at the end of the collection, identical to the collection's `length`.