fix(git/commit): handle inclusive spans without `^`
Previously an inclusive span of commits would use the format `first^..second` for `git log`, which looks for the *parent* of `first`. If `first` is present but its parent is not (for example after a shallow clone from `first` onward) then this function would fail. This behavior was not desired and caused issues generating release notes in PRs, which rely on shallow clones.

A ramification of this fix is that `first` is always included in the output even if the span is nonsensical, for example `second..first`. This new behavior roughly corresponds to the expected behavior from the associated git commands, which do not throw on that nonsensical span so long as `first` and `second` are both valid revisions. Further refinement may be required in the future if this new behavior proves confusing.